Ben First changes

+++ src/server/scripts/Commands/cs_modify.cpp
@@ -1213,8 +1213,9 @@ public:
     {
         if (!*args)
             return false;
-
-        uint16 display_id = (uint16)atoi((char*)args);
+        //BENDEV: Fix uint16 to uint32 to hold bigger display ID's from later expacs
+        // uint16 display_id = (uint16)atoi((char*)args);
+        uint32 display_id = (uint32)atoi((char*)args);
 
         Unit* target = handler->getSelectedUnit();
         if (!target)
